seperating logic for add to cart
